+++ WeatherTweaks/DisplayTable.cs
@@ -17,14 +17,12 @@ namespace WeatherTweaks
 
       var table = new ConsoleTables.ConsoleTable("Planet", "Level weather", "Uncertain weather");
 
-      Plugin.logger.LogWarning($"Displaying weathers table, instance: {StartOfRound.Instance}");
-
       if (StartOfRound.Instance == null)
       {
         return;
       }
 
-      List<SelectableLevel> levels = Variables.GetGameLevels();
+      List<SelectableLevel> levels = MrovLib.LevelHelper.SortedLevels;
       foreach (SelectableLevel level in levels)
       {
         Plugin.logger.LogWarning($"Level: {level.PlanetName}, is null: {level == null}");
@@ -33,7 +31,7 @@ namespace WeatherTweaks
         Plugin.logger.LogDebug($"Is uncertain weather: {isUncertainWeather}");
 
         table.AddRow(
-          level.PlanetName,
+          MrovLib.StringResolver.GetNumberlessName(level),
           Variables.GetPlanetCurrentWeather(level, false),
           isUncertainWeather ? UncertainWeather.uncertainWeathers[level.PlanetName] : ""
         );